Inline branch target resolution, remove BranchTarget class
